# Speed up re-labelling in multi-tenant queries.