--- a/pkg/querier/multi_tenant_querier_test.go
+++ b/pkg/querier/multi_tenant_querier_test.go
@@ -2,10 +2,13 @@ package querier
 
 import (
 	"context"
+	"fmt"
+	"strconv"
 	"testing"
 	"time"
 
 	"github.com/go-kit/log"
+	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/user"
@@ -148,4 +151,52 @@ func newSampleIterator() iter.SampleIterator {
 			StreamHash: labelBar.Hash(),
 		}),
 	})
+}
+
+func BenchmarkTenantEntryIteratorLabels(b *testing.B) {
+	it := NewMockEntryIterator(12)
+	tenantIter := NewTenantEntryIterator(it, "tenant_1") 
+
+	b.ResetTimer()
+	b.ReportAllocs()
+
+	for n := 0; n < b.N; n++ {
+		tenantIter.Labels()
+	}
+}
+
+type mockEntryIterator struct{
+	labels string
+}
+
+func NewMockEntryIterator(numLabels int) mockEntryIterator {
+	builder := labels.NewBuilder(nil)
+	for i := 1; i <= numLabels; i++ {
+		builder.Set(fmt.Sprintf("label_%d", i), strconv.Itoa(i))
+	}
+	return mockEntryIterator{ labels: builder.Labels().String()}
+}
+
+func (it mockEntryIterator) Labels() string {
+	return it.labels
+}
+
+func (it mockEntryIterator) Entry() logproto.Entry {
+	return logproto.Entry{}
+}
+
+func (it mockEntryIterator) Next() bool {
+	return true
+}
+
+func (it mockEntryIterator) StreamHash() uint64 {
+	return 0
+}
+
+func (it mockEntryIterator) Error() error {
+	return nil
+}
+
+func (it mockEntryIterator) Close() error {
+	return nil
 }