**What this PR does / why we need it**:
```
› benchstat before.txt after.txt
name                          old time/op    new time/op    delta
TenantEntryIteratorLabels-16    13.4µs ± 1%     2.4µs ± 1%  -81.81%  (p=0.000 n=10+10)

name                          old alloc/op   new alloc/op   delta
TenantEntryIteratorLabels-16    4.15kB ± 0%    0.90kB ± 0%  -78.40%  (p=0.002 n=8+10)

name                          old allocs/op  new allocs/op  delta
TenantEntryIteratorLabels-16      44.0 ± 0%      29.0 ± 0%  -34.09%  (p=0.000 n=10+10)
```

This patch introduces a cache for re-labeling series from multi-tenant queries.

**Which issue(s) this PR fixes**:
Fixes #5653

**Special notes for your reviewer**:

**Checklist**
- [ ] Documentation added
- [ ] Tests updated
- [ ] Add an entry in the `CHANGELOG.md` about the changes.